backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.4.0
pydantic==2.6.0
Pillow==10.2.0
PyMuPDF==1.23.26
aiofiles==23.2.0
numpy==1.26.0
scikit-learn==1.4.0
nltk==3.8.1
pytesseract==0.3.10
uuid==1.30
